fixed styling errors and added a main star component for everyone to use for overall review
