+++ client/src/components/overview/OverviewStyles.jsx
@@ -9,7 +9,7 @@ const StylesButtons = (props) => {
   // Ensure a default style is selected on page load
   const { stylesArr } = props;
 
-  console.log('stylesArr in OverviewStyles: ', stylesArr);
+  // console.log('stylesArr in OverviewStyles: ', stylesArr);
   // has default? with one set to true and the rest to false
   // has name: for name of style
   // has original_price and sale_price
