+++ client/src/components/reviews/RatingBreakdown.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import StarRating from '../StarRating';
+// import axios from "axios";
+// import styles from '../styleComponents/Reviews.module.css';
+
+const RatingBreakdown = ({ reviews }) => (
+  <StarRating reviews={reviews} />
+);
+
+export default RatingBreakdown;
