+++ client/src/components/reviews/ReviewsList.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import ReviewTemplate from './ReviewTemplate';
-import styles from '../styleComponents/Reviews.module.css';
+// import styles from '../../styleComponents/Reviews.module.css';
 
 const ReviewsList = ({ reviews }) => {
   // const initialReviews = reviews.slice(0, 2);
