+++ client/src/components/reviews/ReviewsSortHeader.jsx
@@ -1,5 +1,5 @@
-import React from "react";
-import axios from "axios";
+import React from 'react';
+// import axios from "axios";
 
 const ReviewsSortHeader = ({ reviews }) => (
   <div>
